Generalize item cards and categories
