+++ app/src/main/java/com/github/k409/fitflow/ui/components/item/Item.kt
@@ -1,16 +1,16 @@
-package com.github.k409.fitflow.ui.screens.inventory
+package com.github.k409.fitflow.ui.components.item
 
 import com.github.k409.fitflow.R
 import com.github.k409.fitflow.model.Fish
 
-internal data class InventoryItem(
+internal data class Item(
     val name: String,
     val description: String,
     val imageResId: Int,
 )
 
 internal val mockDecorations = listOf(
-    InventoryItem(
+    Item(
         "Plant",
         "I don't know what type of plant this is",
         R.drawable.plant,
@@ -18,32 +18,32 @@ internal val mockDecorations = listOf(
 )
 
 internal val mockFishes = listOf(
-    InventoryItem(
+    Item(
         Fish.Primary.title,
         Fish.Primary.description,
         Fish.Primary.getPhaseImage(0.55.toFloat()),
     ),
-    InventoryItem(
+    Item(
         Fish.Secondary.title,
         Fish.Secondary.description,
         Fish.Secondary.getPhaseImage(0.55.toFloat()),
     ),
-    InventoryItem(
+    Item(
         Fish.Third.title,
         Fish.Third.description,
         Fish.Third.getPhaseImage(0.80.toFloat()),
     ),
-    InventoryItem(
+    Item(
         Fish.Third.title,
         Fish.Third.description,
         Fish.Third.getPhaseImage(0.55.toFloat()),
     ),
-    InventoryItem(
+    Item(
         Fish.Third.title,
         Fish.Third.description,
         Fish.Third.getPhaseImage(0.20.toFloat()),
     ),
-    InventoryItem(
+    Item(
         Fish.Third.title,
         Fish.Third.description,
         Fish.Third.getPhaseImage(0.toFloat()),
